UPDATE Function moved on utils
